refactor: dao to repository
