+++ src/main/java/com/yourssu/roomescape/auth/AuthService.java
@@ -1,5 +1,6 @@
 package com.yourssu.roomescape.auth;
 
+import com.yourssu.roomescape.auth.dto.LoginRequest;
 import com.yourssu.roomescape.exception.CustomException;
 import com.yourssu.roomescape.exception.ErrorCode;
 import com.yourssu.roomescape.jwt.TokenProvider;
