+++ src/main/java/com/yourssu/roomescape/auth/dto/CheckLoginResponse.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.auth;
+package com.yourssu.roomescape.auth.dto;
 
 public record CheckLoginResponse(String name) {
 }
