+++ src/main/java/com/yourssu/roomescape/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.auth;
+package com.yourssu.roomescape.auth.dto;
 
 public record LoginRequest(String email, String password) {
 }
