+++ src/main/java/com/yourssu/roomescape/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.member;
+package com.yourssu.roomescape.member.dto;
 
 public class MemberResponse {
     private Long id;
