--- src/main/java/com/yourssu/roomescape/reservation/ReservationWaitingWithRank.java
@@ -1,3 +1,0 @@
-package com.yourssu.roomescape.reservation;
-
-public record ReservationWaitingWithRank(Reservation reservation, Long rank) {}
