+++ src/main/java/com/yourssu/roomescape/reservation/dto/ReservationFindAllResponse.java
@@ -1,3 +1,3 @@
-package com.yourssu.roomescape.reservation;
+package com.yourssu.roomescape.reservation.dto;
 
 public record ReservationFindAllResponse(Long reservationId, String theme, String date, String time, String status) { }
