+++ src/main/java/com/yourssu/roomescape/reservation/dto/ReservationSaveRequest.java
@@ -1,3 +1,5 @@
-package com.yourssu.roomescape.reservation;
+package com.yourssu.roomescape.reservation.dto;
+
+import com.yourssu.roomescape.reservation.ReservationStatus;
 
 public record ReservationSaveRequest(String name, String date, Long theme, Long time, ReservationStatus status) { }
