+++ src/main/java/com/yourssu/roomescape/reservation/dto/ReservationSaveResponse.java
@@ -1,3 +1,5 @@
-package com.yourssu.roomescape.reservation;
+package com.yourssu.roomescape.reservation.dto;
+
+import com.yourssu.roomescape.reservation.ReservationStatus;
 
 public record ReservationSaveResponse(Long id, String name, String theme, String date, String time, ReservationStatus status) {}
