+++ src/main/java/com/yourssu/roomescape/reservation/dto/ReservationWaitingWithRank.java
@@ -1,0 +1,5 @@
+package com.yourssu.roomescape.reservation.dto;
+
+import com.yourssu.roomescape.reservation.Reservation;
+
+public record ReservationWaitingWithRank(Reservation reservation, Long rank) {}
