+++ src/main/java/com/yourssu/roomescape/theme/ThemeController.java
@@ -7,27 +7,29 @@ import java.net.URI;
 import java.util.List;
 
 @RestController
-public class ThemeController {
-    private ThemeDao themeDao;
 
-    public ThemeController(ThemeDao themeDao) {
-        this.themeDao = themeDao;
+public class ThemeController {
+
+    private final ThemeRepository themeRepository;
+
+    public ThemeController(ThemeRepository themeRepository) {
+        this.themeRepository = themeRepository;
     }
 
     @PostMapping("/themes")
     public ResponseEntity<Theme> createTheme(@RequestBody Theme theme) {
-        Theme newTheme = themeDao.save(theme);
+        Theme newTheme = themeRepository.save(theme);
         return ResponseEntity.created(URI.create("/themes/" + newTheme.getId())).body(newTheme);
     }
 
     @GetMapping("/themes")
     public ResponseEntity<List<Theme>> list() {
-        return ResponseEntity.ok(themeDao.findAll());
+        return ResponseEntity.ok(themeRepository.findAll());
     }
 
     @DeleteMapping("/themes/{id}")
     public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeDao.deleteById(id);
+        themeRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 }
