--- src/main/java/com/yourssu/roomescape/time/TimeDao.java
@@ -1,47 +1,0 @@
-package com.yourssu.roomescape.time;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-
-import java.sql.PreparedStatement;
-import java.util.List;
-
-@Repository
-public class TimeDao {
-    private final JdbcTemplate jdbcTemplate;
-
-    public TimeDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Time> findAll() {
-        return jdbcTemplate.query(
-                "SELECT * FROM time WHERE deleted = false",
-                (rs, rowNum) -> new Time(
-                        rs.getLong("id"),
-                        rs.getString("time_value")));
-    }
-
-    public Time findById(Long id) {
-        return jdbcTemplate.queryForObject("SELECT * FROM time WHERE id = ?",
-                (rs, rowNum) -> new Time(rs.getLong("id"), rs.getString("time_value")),
-                id);
-    }
-
-    public Time save(Time time) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        this.jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement("INSERT INTO time(time_value) VALUES (?)", new String[]{"id"});
-            ps.setString(1, time.getValue());
-            return ps;
-        }, keyHolder);
-
-        return new Time(keyHolder.getKey().longValue(), time.getValue());
-    }
-
-    public void deleteById(Long id) {
-        jdbcTemplate.update("UPDATE time SET deleted = true WHERE id = ?", id);
-    }
-}
